added cmake files
> created cmake build files
> shifted libraries around (the libraries will still have to be installed manually for the cmake linux build, they are just here for the windows Visual Studio build)
> changed code to fit C++17's requirements
> cmake build has to be done in SpaceShooter/SpaceShooter/ (unintuitive, I know)

+++ SpaceShooter/ProjectileManager.cpp
@@ -33,7 +33,7 @@ void ProjectileManager::addProjectile(Projectile* p)
 
 void ProjectileManager::removeProjectile(Projectile* p)
 {
-	for (int i = 0; i < projectiles.size(); i++)
+	for (Uint8 i = 0; i < projectiles.size(); i++)
 	{
 		if (projectiles[i] == p)
 		{
